refactor: BookClubRecruit와 Meeting 위치 이동
BookClubRecruit와 Meeting이 domain 밖에 있는 것을 발견하여 이를 수정함.

related to: #8
